third_party/blink/web_tests/external/wpt/html/semantics/interactive-elements/the-details-element/toggleEvent-expected.txt

This is a testharness.js-based test.
FAIL Adding open to 'details' should fire a toggle event at the 'details' element, with 'oldState: closed' and 'newState: open' assert_equals: expected (string) "closed" but got (undefined) undefined
FAIL Removing open from 'details' should fire a toggle event at the 'details' element, with 'oldState: open' and 'newState: closed' assert_equals: Prototype of toggle event is ToggleEvent.prototype expected object "[object ToggleEvent]" but got object "[object Event]"
FAIL Adding open to 'details' (display:none) should fire a toggle event at the 'details' element, with 'oldState: closed' and 'newState: open' assert_equals: expected (string) "closed" but got (undefined) undefined
FAIL Adding open to 'details' (no children) should fire a toggle event at the 'details' element, with 'oldState: closed' and 'newState: open' assert_equals: expected (string) "closed" but got (undefined) undefined
PASS Calling open twice on 'details' fires only one toggle event, with 'oldState: closed' and 'newState: open'
FAIL Calling setAttribute('open', '') from 'details' should fire a toggle event at the 'details' element, with 'oldState: closed' and 'newState: open' assert_equals: expected (string) "closed" but got (undefined) undefined
FAIL Calling removeAttribute('open') from 'details' should fire a toggle event at the 'details' element, with 'oldState: open' and 'newState: closed' assert_equals: Prototype of toggle event is ToggleEvent.prototype expected object "[object ToggleEvent]" but got object "[object Event]"
PASS Setting open=true to opened 'details' element should not fire a toggle event at the 'details' element
PASS Setting open=false to closed 'details' element should not fire a toggle event at the 'details' element
FAIL Adding open to 'details' (not in the document) should fire a toggle event at the 'details' element, with 'oldState: closed' and 'newState: open' assert_equals: Prototype of toggle event is ToggleEvent.prototype expected object "[object ToggleEvent]" but got object "[object Event]"
Harness: the test ran to completion.
